Phase 1 milestone: all 37 kernel C files compile to x86_64 ELF
Warnings remain (92 on a clean rebuild, mostly pointer-to-int and
long*<->int* mismatches that surface real 32-bit assumptions in user
code). All hard errors gone; PortPlan.md M1 milestone hit.

Highlights:

  * h/global.h: ATOX/UPTR/SETTYPE family use uintptr_t casts now.
  * ffasl.c: gated behind !linux_x86_64 -- foreign-function FASL is
    deferred (Phase 6+) and the file would otherwise drag in a.out
    structs we don't have.
  * fex3.c: Linux has no vadvise(2); pagerand/pageseql/pagenorm are
    stubs. Could become madvise calls if hinting becomes useful.
  * fex4.c: <time.h> is required on glibc for struct tm (sys/time.h
    only declares struct timeval).
  * fpipe.c, lam1.c: BSD FILE-internal pokes (_p, _r, _w, _bf, etc.)
    don't compile on glibc; rewritten to use fmemopen, ftello, and
    fflush. Drained-port handling on a TTY is dropped (the kernel
    queue drain was a separate concern from the stdio buffer drain
    and string ports don't need it).
  * lam9.c: Ltci uses termios + cfgetospeed instead of struct sgttyb.
  * inits.c, sysat.c, lam9.c, fpipe.c: forward declarations added at
    file scope to fix "static declaration follows non-static" errors.
  * lam7.c: signal handler typed as SIGTYPE(*)(int) rather than the
    K&R int (*)() that conflicted with glibc's signal(2).
  * lamgc.c: bare "#ifdef" with no symbol changed to "#ifdef METER".
  * lamp.c: int mcntp = (int) mcnts -> intptr_t to keep 64-bit
    pointers from truncating.
  * rlc.c: stubbed (HOLE variant of rawlisp doesn't apply on Linux).
  * io.c: forward decl of finatom/calcnum in dopow() so the K&R-era
    implicit prototype doesn't conflict with the real definition.
  * fasl.c was already fixed in the prior commit.

The Phase 1 work items 1g (asm helpers, callg.c) and 1b (mmap-based
heap layout / OFFSET as runtime base) are still ahead.

+++ franz/h/global.h
@@ -128,7 +128,12 @@ extern long *sp(), stack(), unstack();
 #endif
 
 extern char typetable[];  /*  the table with types for each page  */
-#define ATOX(a1)	((((int)(a1)) - OFFSET) >> 9)
+/* ATOX: pointer -> page index in typetable[]. Each Lisp page is 512
+ * bytes, so the shift is 9. The (uintptr_t) cast is required on
+ * x86_64 -- (int) would truncate the pointer. OFFSET is a runtime
+ * heap base on Linux (Phase 1b will replace the constant 0).
+ */
+#define ATOX(a1)	((((uintptr_t)(a1)) - OFFSET) >> 9)
 #define	TYPE(a1)	((typetable+1)[ATOX(a1)])
 #define	TYPL(a1)	((typetable+1)[ATOX(a1)])
 #define SETTYPE(a1,b,c)   {if((itemp = ATOX(a1)) >= fakettsize) \
@@ -144,7 +149,10 @@ extern char typetable[];  /*  the table with types for each page  */
 #define	HUNKP(a1)	((TYPE(a1) >= 11) & (TYPE(a1) <= 17))
 #define	HUNKSIZE(a1)	((TYPE(a1)+5) & 15)
 
-#define UPTR(x)	((unsigned)(((long)(x))-(long)CNIL))
+/* UPTR: unsigned pointer offset from CNIL, used as a one-shot range
+ * check against datalim in VALID(). 64-bit safe via uintptr_t.
+ */
+#define UPTR(x)	((uintptr_t)(((uintptr_t)(x))-(uintptr_t)CNIL))
 #define VALID(a)	(UPTR(a) <= UPTR(datalim))
 
 #define Popframe() (errp->olderrp)
